trunk: * An attempt to address the cooldown/exhaustion issue (haven't tried to compile it or test it, but hopefully it will compile, and if it compiles it should work).
branches/0.2:
* Fixed compiler warnings.


git-svn-id: file:///usr/local/src/otland/tfssvnrepo/branches/0.2@4035 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ mounts.cpp
@@ -43,7 +43,7 @@ bool Mount::isTamed(Player* player) const
 	if(!player->getStorageValue(PSTRG_MOUNTS_RANGE_START + (tmpId / 31), value))
 		return false;
 
-	int32_t tmp = pow(2, tmpId % 31);
+	int32_t tmp = (int32_t)pow(2, tmpId % 31);
 	return (tmp & value) == tmp;
 }
